requirements.txt
click>=8.0.0
gitpython>=3.1.0
PyYAML>=6.0